Turn all variables denoting the number of unreachable nodes into size_t
This is an extension of Oleg's 3912b34822

+++ sys/net/routing/aodvv2/routingtable.c
@@ -99,7 +99,7 @@ void routingtable_delete_entry(struct netaddr *addr, aodvv2_metric_t metricType)
     }
 }
 
-void routingtable_break_and_get_all_hopping_over(struct netaddr *hop, struct unreachable_node unreachable_nodes[], int *len)
+void routingtable_break_and_get_all_hopping_over(struct netaddr *hop, struct unreachable_node unreachable_nodes[], size_t *len)
 {
     *len = 0; /* to be sure */
